requirements.txt
SQLAlchemy==2.0.30
python-dotenv==1.0.1
streamlit==1.42.0
streamlit-authenticator==0.4.1
python-dotenv==1.0.1
psycopg2==2.9.5
pillow==10.3.0
pandas==2.2.2

## Providers
google-genai==1.16.1
openai==1.55.3
cohere==5.3.2
mistralai==1.7.0
groq==0.25.0